+++ src/Phan/Analysis/PostOrderAnalysisVisitor.php
@@ -2048,6 +2048,7 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
      */
     public function visitNew(Node $node) : Context
     {
+        $class_list = [];
         try {
             $context_node = new ContextNode(
                 $this->code_base,
@@ -2106,7 +2107,9 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
         } catch (Exception $_) {
             // If we can't figure out what kind of a call
             // this is, don't worry about it
-            return $this->context;
+        }
+        if ($this->isInNoOpPosition($node)) {
+            $this->warnNoopNew($node, $class_list);
         }
 
         return $this->context;
@@ -3840,14 +3843,26 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
         // TODO also check AST_UNPACK against variadic
         $arg_names = [];
         foreach ($argument_list_node->children as $i => $arg) {
-            if (!$arg instanceof Node || $arg->kind !== ast\AST_VAR) {
+            if (!$arg instanceof Node) {
+                return;
+            }
+            $is_unpack = false;
+            if ($arg->kind === ast\AST_UNPACK) {
+                $arg = $arg->children['expr'];
+                if (!$arg instanceof Node) {
+                    return;
+                }
+                $is_unpack = true;
+            }
+            if ($arg->kind !== ast\AST_VAR) {
                 return;
             }
             $arg_name = $arg->children['name'];
             if (!\is_string($arg_name)) {
                 return;
             }
-            if ($parameter_list[$i]->getName() !== $arg_name) {
+            $param = $parameter_list[$i];
+            if ($param->getName() !== $arg_name || $param->isVariadic() !== $is_unpack) {
                 return;
             }
             $arg_names[] = $arg_name;
@@ -4084,6 +4099,56 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                 $node->lineno
             );
         }
+    }
+
+    private static function hasEmptyImplementation(Method $method) : bool
+    {
+        if ($method->isAbstract() || $method->isPHPInternal()) {
+            return false;
+        }
+        $stmts = $method->getNode()->children['stmts'] ?? null;
+        if (!$stmts instanceof Node) {
+            // This is abstract or a stub or a magic method
+            return false;
+        }
+        return empty($stmts->children);
+    }
+
+    /**
+     * @param list<Clazz> $class_list
+     */
+    private function warnNoopNew(
+        Node $node,
+        array $class_list
+    ) : void {
+        $has_constructor_or_destructor = count($class_list) === 0;
+        foreach ($class_list as $class) {
+            if ($class->hasMethodWithName($this->code_base, '__construct', true)) {
+                $constructor = $class->getMethodByName($this->code_base, '__construct');
+                if (!$constructor->getPhanFlagsHasState(\Phan\Language\Element\Flags::IS_FAKE_CONSTRUCTOR)) {
+                    if (!self::hasEmptyImplementation($constructor)) {
+                        $has_constructor_or_destructor = true;
+                        break;
+                    }
+                }
+            }
+            if ($class->hasMethodWithName($this->code_base, '__destruct', true)) {
+                $destructor = $class->getMethodByName($this->code_base, '__destruct');
+                if (!self::hasEmptyImplementation($destructor)) {
+                    $has_constructor_or_destructor = true;
+                    break;
+                }
+            }
+            if (!$class->isClass() || $class->isAbstract()) {
+                $has_constructor_or_destructor = true;
+                break;
+            }
+        }
+        $this->emitIssue(
+            $has_constructor_or_destructor ? Issue::NoopNew : Issue::NoopNewNoSideEffects,
+            $node->lineno,
+            ASTReverter::toShortString($node)
+        );
     }
 
     const LOOP_SCOPE_KINDS = [