+++ tests/files/src/0073_splfileobject.php
@@ -1,2 +1,3 @@
 <?php
 new SplFileObject('string');
+// TODO: Build a list of standard library classes that are documented to throw exceptions for parameters with valid types.