+++ tests/files/src/0802_min_max_explode_edge_case.php
@@ -1,0 +1,10 @@
+<?php
+namespace NS802;
+
+function recurs(...$ion) {
+    implode(...$ion);
+    min(...$ion);
+    if (max(...$ion)) {
+        recurs(...$ion);
+    }
+}