+++ tests/files/src/0803_noop_new.php
@@ -1,0 +1,6 @@
+<?php
+class C803 {}
+
+for (new stdClass(); ; new ArrayObject([3])) {
+}
+new C803();